create a new fixture for external renference
