+++ src/ensembl/tests/fixtures/entity-viewer/product.ts
@@ -16,15 +16,14 @@
 
 import faker from 'faker';
 import times from 'lodash/times';
-import merge from 'lodash/merge';
+
+import { createExternalReference } from './external-references';
 
 import {
   ProteinDomain,
   Product,
   ProductType
 } from 'src/shared/types/thoas/product';
-
-import { ExternalReference } from 'src/shared/types/thoas/externalReference';
 
 export const createProduct = (fragment: Partial<Product> = {}): Product => {
   const length =
@@ -43,28 +42,6 @@ export const createProduct = (fragment: Partial<Product> = {}): Product => {
     external_references: times(2, () => createExternalReference()),
     ...fragment
   };
-};
-
-export const createExternalReference = (
-  fragment?: Partial<
-    Omit<ExternalReference, 'source'> & {
-      source: Partial<ExternalReference['source']>;
-    }
-  >
-): ExternalReference => {
-  const xref = {
-    accession_id: faker.lorem.word(),
-    name: faker.lorem.word(),
-    description: faker.lorem.word(),
-    url: faker.lorem.word(),
-    source: {
-      name: faker.lorem.word(),
-      id: faker.lorem.word(),
-      url: faker.lorem.word()
-    }
-  };
-
-  return merge(xref, fragment);
 };
 
 const createProteinDomains = (proteinLength: number): ProteinDomain[] => {
